[TwigBridge][Form] fix tests by moving AbstractBootstrap* cases to the bridge
